benchmark: extracted TLS server script
Refactored out the TLS server into a script that can be run on it's own.

Modified tls-connect benchmark to fork a new process for the TLS server,
so that the store server in master process does not interfere with the
benchmarking client.

Added more configuration options to the benchmark:
- protocol (default, SSLv3_method, etc.)
- number of cluster workers

Simplified the default combination of options.